+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to Scrawler will be documented in this file.
 - Added literal client configuration provider
 - Added total time displayed once operation is done
 - Added ability to set maximum number of crawled URLs
+- Added a warning when empty response is encountered
 - Added ability to mark objects as fetched only once per operation
 - Added ability to pass encoding options to JSON file result writer
 - Added support for relative URL in `ArgumentAdvancerUrlListProvider`'s template